+++ lib/mercury_client.dart
@@ -1,3 +1,6 @@
 library mercury_client;
 
 export 'src/http_client.dart';
+
+export 'src/http_cache.dart';
+