+++ lib/src/http_client_none.dart
@@ -6,7 +6,7 @@ class HttpClientRequesterNone extends HttpClientRequester {
 
   @override
   Future<HttpResponse> doHttpRequest(HttpClient client, HttpRequest request) {
-    return Future.error( HttpError( request.requestURL, 0, 'No HttpClientRequester for ${ request.method } request: ${ request.requestURL }', null) ) ;
+    return Future.error( HttpError( request.url, request.requestURL, 0, 'No HttpClientRequester for ${ request.method } request: ${ request.requestURL }', null) ) ;
   }
 
 }