+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.6
+
+- Public method buildRequestURL.
+- HttpCache class: handles requests cache with entries timeout and memory limit.
+
+
 ## 1.0.5
 
 - Update dependencies.